Rename lambda from hard-disk to copier.
