+++ packages/server/lambdas/src/copier/README.md
