+++ packages/server/lambdas/src/index.ts
@@ -7,6 +7,6 @@ export * from "./broadcaster";
 export * from "./deli";
 export * from "./scribe";
 export * from "./scriptorium";
-export * from "./hard-disk";
+export * from "./copier";
 export * from "./sequencedLambda";
 export * from "./foreman";
