+++ packages/server/routerlicious/package.json
@@ -20,8 +20,8 @@
     "paparazzi:debug": "node --inspect=0.0.0.0:5858 dist/paparazzi/index.js",
     "scriptorium": "node dist/scriptorium/index.js",
     "scriptorium:debug": "node --inspect=0.0.0.0:5858 dist/scriptorium/index.js",
-    "hard-disk": "node dist/hard-disk/index.js",
-    "hard-disk:debug": "node --inspect=0.0.0.0:5858 dist/hard-disk/index.js",
+    "copier": "node dist/copier/index.js",
+    "copier:debug": "node --inspect=0.0.0.0:5858 dist/copier/index.js",
     "foreman": "node dist/foreman/index.js",
     "foreman:debug": "node --inspect=0.0.0.0:5858 dist/foreman/index.js",
     "test": "mocha --recursive dist/test",
